Stop testing the symlink.
The main issue around the symlink is that node is testing not ONE file but TWO files the actual link itself and possibly where that symlink WAS symlinked to. When you restart Local that symlink might point to an OLD file in Local which no longer exists, so the actual link may exist, but the file it points to probably wont.

This just goes back to what we used to do which is try and unlink (delete) the symlink, and if it fails or passes try and create the symlink.

If we create the symlink great, we continue, if not then we can't modify the link period and show a red button...